A major limitation of the previous instruction set was that there was no easy way to do
   v1 = s1

...which is pretty useful  The remedy is to allow vector immediate operations to have scalar operands.  Copy
can then be emulated by using a logical or with zero.  This consumes a bit of the immediate operand field,
which is now only 8 bits.  To improve the situation, the ISA now supports using the mask field as part
of the immediate operand if there is no mask specified (a common situation, and always the case for
scalar fields).

Unfortunately, this breaks the parallelism between A & B op types.  B is supposed to just be
the same as A except with an immediate operand, but now it supports instructions that A doesn't.  The
A format is currently out of bits, so some rearranging of other types would be necessary to make this work.