Fix a bug where long latency instructions would not be rolled back if an earlier instruction missed the data cache.  For example:
  s0 = mem_l[s1]
  s2 = s3 * s4

If the first instruction missed the cache, the multiply would still write back to s2 before the
cache miss could be handled.
Note that this is benign, since [currently] rollbacks initiated after the execute stage will always
be restarted, but it was a little clunky and it broke the verification code because there was a spurious
write that wasn't in the model.
Make the rollback stage now track long latency instructions in the execute unit and roll them back.
This code is a little tricky because the execute stage can reorder instructions.