Cleaned up dcache control signals.
Previously, there was a dcache_request signal, which was qualified with a dcache_store signal (the latter
being false indicated a load).  However, this broke down a bit when the store buffer was moved out of the
dcache, and when dcache_flush and dcache_barrier signals were added (dcache_request is not asserted when the
last two are used)

Now, there are four mutually exclusive signals: dcache_load, dcache_store, dcache_flush, and dcache_barrier.