Fix a bug found in randomizer/regression/test0007.hex
The following scenario:

Thread A starts a vector transfer and loads a few lanes
Thread B misses a cache line x
Thread A misses the same cache line x as part of its vector transfer in the same cycle the fill for B's miss
 completes.

In this case Thread A needs to be restarted.  However, the bug was that it would restart from the first
lane instead of where it left off.  The problem is that there are three rollback scenarios:
1. Mispredicted branch - if a load transfer started, we would want to cancel it and reset the lane to the first
2. Cache miss - Want to suspend the strand and save the last lane that was transferred
3. Retry - Hit hazard where the fill of a cache line occured in the same cycle it was missed by another strand.

Previously, case 3 was not explicitly encoded, os it was treated as case #1.  Added a new signal 'retry' that
explicitly indiciates case 3.