Previously, when a response came from the L2 cache, either the load miss queue or store buffer would match it to a pending request and then transmit the way and tag (which it had latched when the request was made) to the L1 cache to update it.
I realized storing and transmitting the way was totally redundant. There was a mux to select between the way stored in the response and that latched, even though the way from the L2 cache would always be the same as the latched version in situations where the latter should be used.  The set index was previously not contained in the L2 response, but it would be needed in order to perform SMP updates anyway.

I've now modified the code so the L1 cache can directly interpret the L2 response without intervention of the load miss queue or store buffer.  The latter two still keep track of blocked strands and are responsible for waking them up. I added the full address of the request to the response so both the set and tag can be derived.