Instruction set cleanup
- Remove options to invert the mask. These end up being used fairly rarely,
and even in the case that they do, don't win much performance over executing
a few functions to invert the mask. Getting rid of this opens up the
instruction encoding space and simplifies the logic

- Remove strided vector transfer modes. In practice, the strides are often
dynamic. Having the stride hardcoded into the instruction didn't turn out
to be very useful. In most places, it's not too much more work to use scatter
gather transfers to perform the same function.  This may be resurrected in a
different form later.