render test: fix teapot
The teapot was getting clipped because it Z was less than 1.0.  I had
made it be very close up so it would be big enough.  Added a scale
matrix to make it larger when it is inside the view volume.  This also
necessitated normalizing the normal vector inside the pixel shader
because the scale matrix had also scaled it and made it look very
cel shaded.