Fix render tests
Need to increase the amount of testbench memory to 16MB for this to run.
The memory size was previously too small, but verilator seemed to work
correctly when running off the end of the memory array somehow. My previous
change that made it configure the SDRAM correctly based on the memory size
caused this to become apparent.