Fix erroneous assertion triggered when no instruction is fetched
When no threads are ready to fetch an instruction (waiting on icache
hit), the icache tag is not fetched (SRAM read enable is qualified),
but valid is still fetched for which ever PC would be fetched next.
This can result in way_hit_oh having multiple bits set. This is generally
okay, since all of the signals are ignored if no instruction is fetched.
However, the assertion didn't check this bit.