ps2 controller: when overrun occurs, drop oldest character instead of newest
In hardware, when rendering is really slow and a key is repeating, the PS/2
input FIFO can fill. Because it dropped the newest characters, this would lose
the key up scancode, so the key would be stuck down.  Made it drop the oldest
instead, which should behave better.

Also increased FIFO length to 16.