Control thread enable with memory mapped registers instead of control registers
Writing 1 bits to address 0xffff0060 enables the corresponding threads
Writing 1s to 0xffff0064 halts them
This scheme avoids race conditions when threads attempt to halt themselves.

This change addresses several problems.

As all cores share the memory mapped register (as opposed to the control
register, which is private to each core), The system now boots with only one
thread active on core 0 (instead of thread 0 active on all cores), which can
start the other threads.

This makes it easier to scale the number of threads past 32 without
using up limited control register space.

As more interrupt control registers are added, it make sense to have them
all in the same place. Memory mapped registers are the logical place for
these. Halt is strongly tied to interrupts.