hardware: fix Design Compiler synthesis crash
Using ternary operator inside a wire width declaration causes an internal
error. I did this because the width of core_id_t was based on the number
of cores. This was computed by making the high bit be $clog2(NUM_CORES) - 1.
However, this computation does not work for one core.

My alternative is to hard code the core_id_t width (limiting the number of
cores, unless this is changed), and adding special case generator code
to check for one core in the L2 and I/O arbiters.