fix MMU bug with multiple threads
The MMU enable flag was checked against a single that was one cycle early.
In a situation where one thread had the MMU enabled and another didn't,
the one that didn't have it enabled could end up with ift_tlb_hit being
0.  However, TLB miss would not be set in the second stage, nor would
cache_hit or ifd_cache_miss.  The instruction would be thrown away but
the PC would still advance.

Use the registered thread ID to check the MMU enable flag. Added some comments
to try to clarify the behavior.