render tests: update for new compiler
The new version of the compiler has changed the rendered output of
two tests slightly. However, it actually seems to have resolved some
issues with artifacts around the edges of primitives. I haven't been
able to isolate exactly what has changed, so I'm pushing as is.

Also, created a new test texture that is easier to immediately
identify as correct or not.