hardware: Rework L2 dflush to complete memory writeback responding
Previously, the dflush response would be sent after the request
reached the end of the L2 pipeline. I have reworked it so, if a
writeback is required, the response is not sent but is instead
restarted after the writeback completes (in the same manner that
a L2 miss is restarted). This is more complicated, but solves a
few important problems:

- Previously, if a membar was executed after a dflush, it didn't
guarantee the cached values had actually been pushed back to memory.
If this instruction was used to maintain consistency with an external
device (DMA controller or another processor), there would be a race
condition that was difficult to avoid.
- A single thread issuing flushes could starve all other processors
(because this is write through, it could actually make them all stop
executing). The new system rate limits a thread invoking dflushes more
effectively.