Rename some identifiers to be clearer
Previously, I was using the term 'fault' for any asynchronous transfer
of execution, including unaligned data accesses, system calls,
permissions errors, and interrupts. This was a bit misleading, because
fault implies a program error, which not all of these are.

Now, I use the term 'trap' for any asynchronous transfer of execution.
Fault is only for errors.