hardware: Performance counter improvements
- Add new performance events: unconditional branch, conditional branch taken
and not taken
- Fix performance counter test: it was optimizing away the test function.
- Fix a bug where is_eret could be set in int_execute_stage even if a rollback
occurred. If the instruction before eret faults, it could cause the control
registers to be restored as if the eret executed, even though it hasn't.
- Do not need to check if the instrution is illegal or has a piggybacked fault
  in branch logic, since the decoder will not set is_branch in those cases.
- Modify instruction_decode_stage to not set is_legal_instruction if an
interrupt is piggybacked.

A big problem is that some signals like is_valid are ambiguous. They can be unset if
the instruction is valid but is being rolled back, and set if the instruction is
not a real instruction but a piggybacked fault or interrupt. These often feed into
other signals that pick up even more qualifiers. Need to think about a design
convention that simplifies this and makes it more clear.