Fix issue #7 Randomly generated cosimulation tests fail with interrupts enabled
Hardware does not create store cycles for scatter store lanes that do not have
the mask bit set (and thus doesn't emit cosimulation events). It's possible
for the instruction to finish in hardware (because the remaining lanes aren't
enabled in the mask register), but the emulator not to have finished it because
it didn't receive any cosimulation events. When the emulator started executing
the ISR, the subcycle counter would be non-zero, which messed up things later.

The fix is to check explicity if the PC has advanced in the interrupt handler
and reset the subcycle counter.