hardware: add new interface type for io_bus
Also, modified semantics to be more consistent. Previously peripherals asserted
io_read_data asynchrnously, but nyuzi expected it to have a cycle of latency.
The top level module added a flop to synchronize it. Now, everybody who uses
an io interface expects it to have a cycle of latency. In the top level module,
it latches the source in the first cycle and uses an async mux to select the
source in the second. This probably adds a little delay, but is cleaner.