Fix verilator submodule
I was using checkout -b, which was creating a new branch from master
instead of checking out the appropriate tag.
Adjusted the subproject to point to the comment for tag verilator_3_880.
Removed the command from setup_tools.sh to check out the branch, since
it is now pointing to the proper commit.

[ci skip]