Check 'present' TLB bit in hardware
Previously, I attempted to do this in software in the TLB miss handler.
However, it seems impossible using the current instruction set to dispatch a
page fault exception from the TLB miss handler. It would need to jump to the
page fault virtual address and reenable the MMU in the same cycle. The only
instruction that can do that is an eret, but doing so would clobber the reset
of the exception state in control registers, especially with a nested TLB miss.

While this has slightly more overhead when a page fault occurs (because it will
usually take both a TLB miss, then a page fault exception), it is faster in the
common case where there isn't a page fault because the TLB miss handler doens't
need to explicitly check the present bit.