hardware: change interrupt/load_sync logic
Previously this would handle an interrupt in the middle of load_sync
by resetting the state. Instead I now disallow an interrupt to be
dispatched when a load_sync is pending. This is more consistent with
how other multi-issue instructions (like I/O read/writes) work,
and enables the next set of interrupt changes.