hardware: change how interrupts are identified
This is another step towards decoupling the interrupt controller and
execution pipeline. Instead of the interrupt number being encoded
in the trap reason control register, there is a single trap reason
for all interrupts. My plan is to add an I/O register in the interrupt
controller for identifying which external device raised the interrupt.