Add more information to trap cause control register
Previously, this was a reason register that had a single integer in it.
I have modified it to have a 4 bit type code, and two extra bits to
indicate if a memory access was a load or store, and whether the
instruction or data access caused the fault.

The main reason for this is to determine if a page fault is caused by
a store or load, which simplifies copy-on-write and dirty bit handling.

This also fixes a test bug: the instruction_supervisor_fault MMU test
was actually causing a page fault and not checking that the proper
fault type occurred.