emulator: fix bug when interrupts occurred during multi-cycle instructions
This fixes a crash when running the plasma demo under the kernel

Normally thread->pc points to the next instruction. However, this is not the
case when a multicycle instruction like load_gath is executing. The trap
handler was subtracting four to record the trap PC, but this was the wrong
thing to do in the load_gath case because it would repeat the previous instruction.