kernel: fix issue #47: KERNEL PANIC: ASSERT FAILED: rwlock.c:75: m->active_read_count > 0
handle_page_fault would always lock the user address space for the current
user thread, but if the fault occured in kernel space, it would unlock
the kernel address space. The problem was that the space variable was
being set twice: once at the beginning using the current thread, then again
based on the address. Fix code so it is set once and the same space is used
for lock and unlock.

The fact that a page fault is occurring in kernel space is a different bug.