kernel: fix deadlock in create_area syscall
I was being a bit lazy and didn't do a safe usercopy of the area name.
The problem is that the name is copied into the area struct with the address space
lock held as a writer. If a valid page fault occurred (page not present, but
address is in an valid area), this would deadlock as the fault handler attempted
to acquire the address space lock as a reader.

The fix was to add a safe user_strlcpy and call this to copy into a temporary
stack array before passing to create_area.