Fix bugs with cache ref counting
vm_areas were not incrementing the cache ref count when
attached to it. This was masked because the caches were created
with a ref count of 1, and they were only mapped once because of
the structure of the program. When linking with LLD, there's an
extra segment, which would result in an assertion because
they would be decremented twice. A few fixes:

- vm_caches are still created with a ref count of 1, but load_program
now releases the ref when it is finished (since the newly created
areas will now hold their own refs)
- Made create_area increment the ref for the area.

Also: there was a check in the loader if there were too many segments,
but it was after the segments were loaded (and had clobbered the program
header table if there were too many). Put the check first, and also add
a check if there are no segments.