emulator: fix issues with remote GDB read/write register
Move parse_hex_vector out of cosimulation and use for remote-gdb. This
exposed a few problems:
- When reading/writing vector registers (which start at register index 32),
wasn't subtracting 32. This was running past the end of the register array
and reading/writing random memory.
- The order of vector lanes was reversed in these cases
- Added a test program that pre-loads values into vector and scalar registers
so I can confirm those values are read correctly in test_read_write_register.
Previously, it was setting values itself, which masked the issues above