Fix two issues when running with multiple cores
- cr_interrupt_pending was too wide. It only needs to have signals for threads
on its own core. This would cause a warning, but had no functional impact.
- If there were between 2 and 8 cores configured, the high bits of grant_idx
in io_interconnect would be unassigned (and undefined). This could result
in an IO response packet being send with an invalid core ID (>8) because
Verilator is configured to set uninitalized bits to random values. This in
turn would cause the processor to hang on an I/O memory read/write because
it would never match the response. Added code to initialize these bits.