python tests: add lint checker targets and fix warnings
Added two new targets for pyflakes and pylint, which will report warnings
for sources. These are useful for catching issues that are usually only
reported at runtime by python.

This required switching away from using a wildcard import of test_harness,
so I needed to qualify all of the things that are used from there now. That
is apparently proper Python practice anyway, but makes the scripts more wordy.

Fixed a bunch of warnings, including but not limited to
- Remove unneeded imports
- Use _ for parameters and unpack variables that are unused
- Convert module and function comments to docstrings
- Make some variable names more descriptive
- Move constants out of local variables and be more consistent with formatting.

Also added target to run autopep8 to reformat sources.