Make any control register read fault in non-privileged mode
The point of this change and the last is to enable virtualization.
Trapping on control register reads allows a hypervisor to subsitute
values for them to emulate hardware.

On further reflection, it's not necessary to make the hardware thread ID
register available from user space, since user programs will want
to find out the OS thread ID. Furthermore, by making this fault on
read, a hypervisor can emulate virtual hardware threads.

Once this is removed, enabling the cycle counter read from
user mode is just an optimization. Since it's only 32 bits, this is
arguably not that useful anyway, so I just eliminated it.

This also simplifies the hardware implementation and testing, since all
registers are treated the same.

Fixed a bug in tests/cosimulation/interrupt.S where it was switching to
user mode in the middle of the test (which became obvious with this change)