Remove the 'ball' and 'bnall' instructions
These would branch in conditions where some or no lanes would execute
in a divergent single-program/multiple-data execution path. However,
they turn out not to be that helpful and are not used.
As it turns out, the only case where we really need to branch is
if the execution mask is zero, which can be done with bfalse.