emulator: Change numbering convention for vector lanes
When I defined the original instruction set, the lane index corresponded
to the bit index in the mask (with 0 as the least significant bit), but
lane 15 was the lowest address when a block load was performed. This
was the opposite of the conventions used by GCC vector extensions
and by LLVM IR. So I modified the instruction set to use the more standard
convention for memory order (lane 0 is the lowest address), but kept
the same mask bit order.

However, I kept the order in the vector value arrays inside the emulator
the same and just added a bunch of conversions throughout the code
to reverse the index. This was bad because:
- Extra code slows stuff down and is crufty
- The 'lane' variables did not match the conventions in documentation and
  the compiler backend.

So I reversed the order of the vector arrays to match the convention
(element 0 of the array is element 0 of the vector), which should be
cleaner and easier to reason about.