Fix failing LLDB test
My changes to the way constants are loaded changed the memory address of the
breakpoint instruction. The test is a bit fragile that it depends on that.

This didn't cause a TravisCI failure because it is not run during continuous
integration, since I don't install LLDB libraries in the build container
(because they're big)

[ci skip]