emulator: Add ability to schedule threads randomly
Previously, the emulator would always run threads in strict round-robin order.
Being entirely deterministic, this doesn't test inter-thread synchronization
well. Added a new flag '-a' that will schedule threads in random order.
This is not on by default because it slows emulation down substantially.

A few other fixes:
- Switch to using a better xorshift128+ random generator instead of the default rand.
It is about 20% slower, but has a longer period and better characteristics.
- The execute_instructions function took a thread_id as a parameter. This could be
set to a special value ALL_THREADS to cycle through the threads. As it turns out,
it was never called without this value, so removed the parameter entirely.

Did a bunch of other minor cleanups.

Total execution time of teapot render test (assuming old performance time was
normalized to 1.0). Lower is better:
With randomization enabled: 1.64x
With randomization disabled: 1.035x

Caveats:
- None of the tests now verify the default round-robin scheduling order,
so if something broke there, CI wouldn't catch it (I noticed this because
there was originally a bug, which I only caught when trying to run
one of the programs in the app directory, which use the default mode).