Add Dhrystone benchmark
In order to get accurate numbers in the emulator, I've modified the emulator
so that the cycle count control register returns the number of instructions
instead of simulating a constant clock rate based on wall clock time. I had
made it work like that to make the clock() libc function accurate in the
emulator, which in turn was to enable FPS counters on apps. I've filed
Bug #118 to cover coming up with an alternative for that use case.