hardware: Refactor trap control register handling
In control_registers module, I now encapsulate all trap state in a
structure. This makes the code more concise and cleaner, but also
allows exploration of a few ideas:
- Currently, these are implemented as a hardware stack that acts
  sort of like a shift register. An alternative might look a little
  more like a ring buffer. This change makes it easier for me to
  explore synthesis tradeoffs.
- I'm considering implementing JTAG support with a special interrupt.
  This change allows much more easily adding another nesting level of
  interrupts (it would have been much more messy before).

I also think I fixed a serious bug: executing eret and taking a trap
were treated as mutually exclusive (there was an 'else if'), but
since these clauses apply to different threads, it seemed to be possible
that if one thread was taking a trap and another thread executed eret
in the same cycle (a few pipeline stages earlier), the eret would not
have the proper side effects.