Fix issue #121: JTAG test intermittently fails
This was a bit of a rabbit hole

The root cause was that the reset block in sim_jtag was not executing. Because
registers are randomized at startup, in some cases need_ir_shift or need_dr_shift
would be set, but shift_count would be some arbitrary value, so it would end up
trying to shift out billions of instruction bits and the test would time out.

The reason the reset block wasn't executing was more interesting. The block
in sim_jtag that wasn't executing looked like this:

    always @(posedge jtag.tck, posedge reset)
    begin
        if (reset)
            // I never get executed...

This looks similar to every other block in the system, except that it is
triggered on the positive edge of jtag.tck instead of the system clock. This
shouldn't be a problem, because this also triggers on the positive edge of reset.
But it turns out, this block was exposed a latent problem that was masked in
the rest of the code. The main simulator loop in verilator_main.cpp looks
like this:

    testbench->reset = 1;
    testbench->clk = 0;

    while (!Verilated::gotFinish())
    {
        if (currentTime > 10)
            testbench->reset = 0;   // Deassert reset

        testbench->clk = !testbench->clk;
        testbench->eval();

Reset starts asserted, then is released after some number of clock cycles. A problem
becomes apparent that there is no rising edge on reset, since it is high in the first
cycle, so this should never work. However, the problem is even more subtle than that.
Internally, verilator maintains a 'last' value for every signal. It compares the current
value to the last to determine if there was an edge. For example, for this chunk of
verilog:

    always @(posedge clk, posedge reset)

Verilator generates the following c++ code:

    if ((((IData)(vlTOPp->clk) & (~ (IData)(vlTOPp->__Vclklast__TOP__clk)))
        | ((IData)(vlTOPp->reset) & (~ (IData)(vlTOPp->__Vclklast__TOP__reset))))) {
        vlSymsp->TOP__soc_tb._sequent__TOP__soc_tb__5(vlSymsp);
    }

Because we are randomizing the values at startup, last is a random value:

void Vsoc_tb::_ctor_var_reset() {
    __Vclklast__TOP__reset = VL_RAND_RESET_I(1);

So if the last was selected as 0, then I manually set reset to 1, tne next eval would
detect a rising edge and the reset clauses would execute:

But why didn't I see this problem until now? The issue was that I was toggling the
clock with reset held. Because clk is also a trigger for most blocks, they would
execute and see reset asserted, which effectively made reset synchronous. But since
jtag.tck is held low while in reset, the block in sim_jtag would never execute.

I've modified verilator_main to start with reset low, then assert and deassert it
before ever toggling the clock. This has the added benefit that it verifies that
all blocks in the design properly handle async reset.