Change reset handling again
With the latest tip-of-tree verilator, the design was asserting all over the place,
because the old reset behavior was somewhat fragile w.r.t. how blocks got
scheduled. Changed to a different hack to more robustly reset the part: by
setting the 'last' value of the reset signal to zero and reset to one, we
ensure reset blocks get triggered at time 0, but the design doesn't trip
asserts at startup before the reset has occurred.