Move scoreboard logic into its own module
This is advantageous because:
- It allows running unit tests on scoreboard logic, which was difficult to
impossible when it was spaghetti inside thread_select_stage
- It allows focused hardware optimization. This module has a lot of logic
and is a good target for this.
- It makes thread_select_stage much cleaner and easier to understand

As part of this change, I've eliminated the extra stage for decoding scoreboard
signals. This will probably have an impact on Fmax, but I believe the original
optimization didn't net a whole lot of improvement. I think this will help
overall performance, since there is less latency to restart instructions after
a rollback.

I had to alter the perf_counters test a bit, because there are fewer icache hits.
This may be due to lower pipeline latency.