test harness: don't run all supported tests by default
There is now a 'DEFAULT_TESTS' variable, which can be overridden by
the --target flag. This makes it possible to add test targets that
don't get run by default when runtest.py is invoked.

- Modified compiler test so there is now a 'host' target, which is not
a default target.

Also change so a number of tests that had hard-coded lists of tests now
find them in the directory automatically. This makes them consistent with
other tests, and also reduces the changes that a test will be accidently
skipped.