Add hardware unit test for dcache_data_stage and major rework
I found a bunch of issues that didn't obviously break functionality, but were
questionable. For example if an access had a permissions fault, it could still
cause a cache refill. This could leak information about the kernel memory map.

There was what looked like a real bug where a gather load could cause a fault
even if the lane was masked. I apparently only tested this with stores. Make
this more generic so it is tested for all memory access types.

The logic in this module in general was pretty hard to follow and spaghetti-ish.
A major problem is that two concepts were much together in the enable signals
(e.g. dcache_load_en): if the instruction is requesting an operation, and if
it should actually do it. I split these explicitly into request signals (_req)
for each operation and enable (_en) signals that perform the operation. In
between is fault checking logic. I also reordered the module strictly so each
signal is assigned before it is read. This is not required by Verilog, but makes
the logic easier to follow.

A few other changes:
- In core/isa/load_store.S, the mask I was using for gather loads was a binary
palindrome, which could mask bugs where the mask bits where interpreted backwards.
- Added --trace-structs flag to unit test script when --debug is enabled.
- Renamed some signals to be clearer.