Issue #62: Add JTAG debugging support to hardware pipeline
Added support for querying status of an injected instruction.

This fixes a big issue, where the debugger couldn't know if
an injected instruction was rolled back (for example, because of a
cache miss).

I also fixed another issue that I hadn't realized yet, which was
that an interrupt could have replaced the injected instruction.